update/ adjust handleJoinSubmit
